refactor: remove CompiledSchema properties from ExecutionContext
